Harden sales return and stock flows

+++ database/migrations/2025_03_10_155508_create_physical_quantities_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->date('date');
             $table->unsignedBigInteger('article_id');
-            $table->integer('packets');
+            $table->decimal('packets', 10, 2);
             $table->string('category');
             $table->timestamps();
